+++ src/cmds/mod.rs
@@ -16,9 +16,6 @@ pub mod grow;
 /// Init command
 pub mod init;
 
-/// Inspect command
-pub mod inspect;
-
 /// List command
 pub mod list;